requirements.txt
Flask
Flask-SQLAlchemy
Werkzeug
PyJWT
psycopg2
python-dotenv
flask-cors